Refactor openrouter-test into tests package and add test modules
Refactor test suite: split monolithic cmd/openrouter-test/main.go test functions into a new tests package under cmd/openrouter-test/tests/. Added helper utilities (helpers.go) and multiple test modules: chat, account, advanced, apikeys, models, routing, search. Updated main.go to import and call tests package functions (replaced local run* functions with tests.Run*). Also: removed unused encoding/json import from main.go. README.md updated to add E2E test docs and unit test section. This groups changes related to reorganizing the test codebase and documentation updates into one focused commit.
