+++ cmd/openrouter-test/tests/routing.go
@@ -1,0 +1,215 @@
+package tests
+
+import (
+	"context"
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/hra42/openrouter-go"
+)
+
+// RunProviderRoutingTest tests provider routing options
+func RunProviderRoutingTest(ctx context.Context, client *openrouter.Client, model string, maxTokens int, verbose bool) bool {
+	fmt.Printf("🔄 Test: Provider Routing\n")
+
+	messages := []openrouter.Message{
+		openrouter.CreateUserMessage("Say 'Hello from provider routing test' and nothing else."),
+	}
+
+	// Test 1: Provider order with fallbacks
+	fmt.Printf("   Testing provider order...\n")
+	start := time.Now()
+	resp, err := client.ChatComplete(ctx, messages,
+		openrouter.WithModel(model),
+		openrouter.WithMaxTokens(maxTokens),
+		openrouter.WithProviderOrder("openai", "anthropic", "google"),
+		openrouter.WithAllowFallbacks(true),
+	)
+	elapsed := time.Since(start)
+
+	if err != nil {
+		fmt.Printf("   ⚠️  Provider order test failed: %v\n", err)
+		// Don't fail entirely as provider might not be available
+	} else {
+		fmt.Printf("   ✅ Provider order (%.2fs)\n", elapsed.Seconds())
+		if verbose {
+			fmt.Printf("      Response: %s\n", strings.TrimSpace(resp.Choices[0].Message.Content.(string)))
+		}
+	}
+
+	// Test 2: Ignore providers
+	fmt.Printf("   Testing ignore providers...\n")
+	start = time.Now()
+	_, err = client.ChatComplete(ctx, messages,
+		openrouter.WithModel(model),
+		openrouter.WithMaxTokens(maxTokens),
+		openrouter.WithIgnoreProviders("cohere"),
+	)
+	elapsed = time.Since(start)
+
+	if err != nil {
+		fmt.Printf("   ⚠️  Ignore providers test failed: %v\n", err)
+	} else {
+		fmt.Printf("   ✅ Ignore providers (%.2fs)\n", elapsed.Seconds())
+	}
+
+	// Test 3: Require parameters
+	fmt.Printf("   Testing require parameters...\n")
+	start = time.Now()
+	_, err = client.ChatComplete(ctx, messages,
+		openrouter.WithModel(model),
+		openrouter.WithMaxTokens(maxTokens),
+		openrouter.WithRequireParameters(true),
+		openrouter.WithTemperature(0.5),
+	)
+	elapsed = time.Since(start)
+
+	if err != nil {
+		fmt.Printf("   ⚠️  Require parameters test failed: %v\n", err)
+	} else {
+		fmt.Printf("   ✅ Require parameters (%.2fs)\n", elapsed.Seconds())
+	}
+
+	printSuccess("Provider routing tests completed")
+	return true
+}
+
+// RunZDRTest tests zero data retention
+func RunZDRTest(ctx context.Context, client *openrouter.Client, model string, maxTokens int, verbose bool) bool {
+	fmt.Printf("🔄 Test: Zero Data Retention (ZDR)\n")
+
+	messages := []openrouter.Message{
+		openrouter.CreateUserMessage("What is 1+1? Reply with just the number."),
+	}
+
+	start := time.Now()
+	resp, err := client.ChatComplete(ctx, messages,
+		openrouter.WithModel(model),
+		openrouter.WithMaxTokens(maxTokens),
+		openrouter.WithZDR(true),
+	)
+	elapsed := time.Since(start)
+
+	if err != nil {
+		// ZDR might not be available for all models/providers
+		fmt.Printf("⚠️  ZDR not available: %v\n", err)
+		fmt.Printf("   (This is expected if the model/provider doesn't support ZDR)\n")
+		return true // Don't fail the test
+	}
+
+	printSuccess(fmt.Sprintf("Success with ZDR enabled! (%.2fs)", elapsed.Seconds()))
+
+	if verbose {
+		fmt.Printf("   Response: %s\n", strings.TrimSpace(resp.Choices[0].Message.Content.(string)))
+		fmt.Printf("   Model: %s\n", resp.Model)
+	}
+
+	return true
+}
+
+// RunModelSuffixTest tests model suffixes like :nitro and :floor
+func RunModelSuffixTest(ctx context.Context, client *openrouter.Client, model string, verbose bool) bool {
+	fmt.Printf("🔄 Test: Model Suffixes (Nitro/Floor)\n")
+
+	messages := []openrouter.Message{
+		openrouter.CreateUserMessage("Reply with 'OK' and nothing else."),
+	}
+
+	// Test 1: Nitro suffix for throughput optimization
+	fmt.Printf("   Testing :nitro suffix...\n")
+	start := time.Now()
+	resp, err := client.ChatComplete(ctx, messages,
+		openrouter.WithModel(model+":nitro"),
+		openrouter.WithMaxTokens(10),
+	)
+	elapsed := time.Since(start)
+
+	if err != nil {
+		fmt.Printf("   ⚠️  Nitro test failed: %v\n", err)
+		// Continue with other tests
+	} else {
+		fmt.Printf("   ✅ Nitro suffix (%.2fs)\n", elapsed.Seconds())
+		if verbose {
+			fmt.Printf("      Response: %s\n", strings.TrimSpace(resp.Choices[0].Message.Content.(string)))
+		}
+	}
+
+	// Test 2: Floor suffix for lowest price
+	fmt.Printf("   Testing :floor suffix...\n")
+	start = time.Now()
+	resp, err = client.ChatComplete(ctx, messages,
+		openrouter.WithModel(model+":floor"),
+		openrouter.WithMaxTokens(10),
+	)
+	elapsed = time.Since(start)
+
+	if err != nil {
+		fmt.Printf("   ⚠️  Floor test failed: %v\n", err)
+	} else {
+		fmt.Printf("   ✅ Floor suffix (%.2fs)\n", elapsed.Seconds())
+		if verbose {
+			fmt.Printf("      Response: %s\n", strings.TrimSpace(resp.Choices[0].Message.Content.(string)))
+		}
+	}
+
+	printSuccess("Model suffix tests completed")
+	return true
+}
+
+// RunPriceConstraintTest tests price constraints and data collection policies
+func RunPriceConstraintTest(ctx context.Context, client *openrouter.Client, model string, maxTokens int, verbose bool) bool {
+	fmt.Printf("🔄 Test: Price Constraints\n")
+
+	messages := []openrouter.Message{
+		openrouter.CreateUserMessage("Say 'Price test OK' and nothing else."),
+	}
+
+	// Test with max price constraints
+	maxPrice := openrouter.MaxPrice{
+		Prompt:     5.0,  // $5 per million prompt tokens
+		Completion: 10.0, // $10 per million completion tokens
+	}
+
+	start := time.Now()
+	resp, err := client.ChatComplete(ctx, messages,
+		openrouter.WithModel(model),
+		openrouter.WithMaxTokens(maxTokens),
+		openrouter.WithMaxPrice(maxPrice),
+		openrouter.WithProviderSort("throughput"), // Get fastest provider within price limit
+	)
+	elapsed := time.Since(start)
+
+	if err != nil {
+		printError("Failed", err)
+		return false
+	}
+
+	printSuccess(fmt.Sprintf("Success with price constraints! (%.2fs)", elapsed.Seconds()))
+	fmt.Printf("   Max price: $%.2f/M prompt, $%.2f/M completion\n", maxPrice.Prompt, maxPrice.Completion)
+
+	if verbose {
+		fmt.Printf("   Response: %s\n", strings.TrimSpace(resp.Choices[0].Message.Content.(string)))
+		fmt.Printf("   Model used: %s\n", resp.Model)
+		fmt.Printf("   Tokens: %d prompt, %d completion\n",
+			resp.Usage.PromptTokens, resp.Usage.CompletionTokens)
+	}
+
+	// Test with data collection policy
+	fmt.Printf("   Testing data collection policy...\n")
+	start = time.Now()
+	_, err = client.ChatComplete(ctx, messages,
+		openrouter.WithModel(model),
+		openrouter.WithMaxTokens(maxTokens),
+		openrouter.WithDataCollection("deny"),
+	)
+	elapsed = time.Since(start)
+
+	if err != nil {
+		fmt.Printf("   ⚠️  Data collection 'deny' not available: %v\n", err)
+	} else {
+		fmt.Printf("   ✅ Data collection policy (%.2fs)\n", elapsed.Seconds())
+	}
+
+	return true
+}
